# [action] [PR:563] [cmis] combine cmis eeprom i2c read