… sonic_platform_base/sonic_xcvr/fields/consts.py