Change multiple eeprom i2c reads into a single i2c read across multiple addresses, as addresses are continuous in register group.
<!-- Provide a general summary of your changes in the Title above -->

#### Description
<!--
     Describe your changes in detail
-->
1. Update functions get_tx_power_flag, get_tx_bias_flag and get_rx_power_flag.
2. Add dedicated groups: TX_POWER_ALARM_FLAGS_FIELD, TX_BIAS_ALARM_FLAGS_FIELD, RX_POWER_ALARM_FLAGS_FIELD.
3. Move the MEDIA_TYPE_FIELD read and prefix lookup out of the loop.

#### Motivation and Context
CMIS API methods were issuing several separate i2c transactions per call, resulting in high bus overhead, increased latency, and poor scalability. By introducing dedicated RegGroupField groups for TX/RX power and bias alarm flags (e.g. TX_POWER_ALARM_FLAGS, TX_BIAS_ALARM_FLAGS, RX_POWER_ALARM_FLAGS) and refactoring the related methods (get_tx_power_flag, get_tx_bias_flag, get_rx_power_flag, etc.) to use a single block read per group, it resolves API call latency. Also moved constant prefix lookups (media‐type prefix) outside of loops to save time.

#### How Has This Been Tested?
Test with performance cases for one port, if there are 64 physical ports, then it can save 1.3s.
Before:
```bash
Running performance tests (3000 iterations each)...
get_tx_power_flag: total 32.120s, avg 10.707ms per call
get_tx_bias_flag: total 26.635s, avg 8.878ms per call
get_rx_power_flag: total 32.306s, avg 10.769ms per call
```
After:
```bash
Running performance tests (3000 iterations each)...
get_tx_power_flag: total 8.213s, avg 2.738ms per call
get_tx_bias_flag: total 7.831s, avg 2.610ms per call
get_rx_power_flag: total 9.632s, avg 3.211ms per call
```
Test with config reload, works fine.