@mssonicbld 2025-05-15
Original PR: https://github.com/sonic-net/sonic-platform-common/pull/563